requirements.txt
aiogram==2.25.1
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
amqp==5.1.1 ; python_version >= '3.6'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
babel==2.9.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
kombu==5.2.4 ; python_version >= '3.7'
magic-filter==1.0.9 ; python_version >= '3.7' and python_version < '4.0'
multidict==6.0.4 ; python_version >= '3.7'
prompt-toolkit==3.0.38 ; python_full_version >= '3.7.0'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
rsa==4.9
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
vine==5.0.0 ; python_version >= '3.6'
wcwidth==0.2.6
yarl==1.9.2 ; python_version >= '3.7'
